net_processing: move PeerManagerImpl into cpp file
